pyproject.toml
[project]
name = "deep-thought"
version = "0.1.0"
description = ""
authors = [
    {name = "William Trekell", email = "wtrekell@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "todoist-api-python>=3.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "marker-pdf>=1.0.0",
    "markitdown>=0.1.0",
    "playwright>=1.40.0",
    "html2text>=2020.1.16",
    "extract-msg>=0.50.0",
    "mlx-whisper>=0.4.0",
    "praw>=7.7.1",
    "google-api-python-client>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.2.0",
    "google-generativeai>=0.5.0",
    "httpx>=0.27.0",
    "rich>=13.0.0",
]

[project.scripts]
todoist = "deep_thought.todoist.cli:main"
file-txt = "deep_thought.file_txt.cli:main"
web = "deep_thought.web.cli:main"
audio = "deep_thought.audio.cli:main"
reddit = "deep_thought.reddit.cli:main"
gmail = "deep_thought.gmail.cli:main"
gcal = "deep_thought.gcal.cli:main"
search = "deep_thought.research.cli:search_main"
research = "deep_thought.research.cli:research_main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "types-pyyaml>=6.0.12.20250915",
]
whisper = [
    "openai-whisper",
]
diarize = [
    "pyannote.audio>=3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
]
ignore = []
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
check_untyped_defs = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "error_handling: marks tests for error handling scenarios",
]

